fix: specify fp32 for softmax in load_balancing_loss_func to avoid fp16 underflow

+++ src/transformers/models/glm4v_moe/modeling_glm4v_moe.py
@@ -1553,7 +1553,7 @@ def load_balancing_loss_func(
         compute_device = gate_logits[0].device
         concatenated_gate_logits = torch.cat([layer_gate.to(compute_device) for layer_gate in gate_logits], dim=0)
 
-    routing_weights = torch.nn.functional.softmax(concatenated_gate_logits, dim=-1)
+    routing_weights = torch.nn.functional.softmax(concatenated_gate_logits, dtype=torch.float, dim=-1)
 
     _, selected_experts = torch.topk(routing_weights, top_k, dim=-1)
